% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\title{Create outpack server}
\usage{
api(root, validate = NULL, log_level = "info")
}
\arguments{
\item{root}{Outpack root}

\item{validate}{Logical, indicating if validation should be done
on responses.  This should be \code{FALSE} in production
environments.  See \link[porcelain:porcelain]{porcelain::porcelain} for details}

\item{log_level}{Logging level to use. Sensible options are "off",
"info" and "all".}
}
\value{
A \link[porcelain:porcelain]{porcelain::porcelain} object. Notably this does \emph{not}
start the server
}
\description{
Create an outpack server, a porcelain object
}
